refactor(epg): separate worker runtime and import lifecycle
